update modularized code
